organize into multiple files; make listing methods show fully-qualified names; address small feedback from recent change (trim then check if empty)
